+++ src/controllers/setup.py
@@ -7,7 +7,7 @@ package_name = 'controllers'
 setup(
     name=package_name,
     version='0.0.0',
-    packages=[package_name],
+    packages=[package_name, f'{package_name}.path_planners'],
     data_files=[
         ('share/ament_index/resource_index/packages',
             ['resource/' + package_name]),
